% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{notebook_info}
\alias{notebook_info}
\title{list notebooks in a quiver library}
\usage{
notebook_info(.x, .df = FALSE)
}
\arguments{
\item{.x}{quiver library}
}
\description{
list notebooks in a quiver library
}
