% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{write_content}
\alias{write_content}
\title{write content template to a file}
\usage{
write_content(.ct, .file, append = FALSE)
}
\arguments{
\item{.ct}{content template}

\item{.file}{output file}

\item{append}{whether to append}
}
\description{
write content template to a file
}
