% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{template_content}
\alias{template_content}
\title{turn cells to markdown}
\usage{
template_content(.content, asset_path = "assets")
}
\arguments{
\item{.content}{content}

\item{asset_path}{relative path to assets to content}
}
\description{
turn cells to markdown
}
